AST: simulate inheritance of Declaration using first member of subclasses
This allows converting subclasses of Declaration back to the base class (and vice versa).

This will be useful for parsing subprograms (some subprograms are top-level compilation units,
some are normal declarations. Can just allocate SubprogramDecl and return it to be used as needed
instead of having to allocate a union of all Declaration subclasses)